feat(header desktop): adds desktop header component
