+++ src/layouts/default.vue
@@ -14,21 +14,7 @@
       <HeaderMobile @toggle-sidebar="toggleSidebar()"></HeaderMobile>
       <main class="flex-1 relative z-0 overflow-y-auto focus:outline-none">
         <!-- Page title & actions -->
-        <div
-          class="
-            border-b border-gray-200
-            px-4
-            py-4
-            sm:flex sm:items-center sm:justify-between sm:px-6
-            lg:px-8
-          "
-        >
-          <div class="flex-1 min-w-0">
-            <h1 class="text-lg font-medium leading-6 text-gray-900 sm:truncate">
-              {{ route.meta.title || 'Caja' }}
-            </h1>
-          </div>
-        </div>
+        <HeaderDesktop></HeaderDesktop>>
 
         <div class="px-4 mt-6 sm:px-6 lg:px-8">
           <router-view></router-view>
@@ -39,10 +25,7 @@
 </template>
 
 <script setup lang="ts">
-  import { useRoute } from 'vue-router'
   import { useSidebarMachine } from '~/state/sidebar-machine'
-
-  const route = useRoute()
 
   const persistedState = JSON.parse(
     localStorage.getItem('user-menu-state') ?? '{}'
